+++ app/src/main/java/com/example/kuitandroidapiexample/MainActivity.kt
@@ -7,7 +7,7 @@ import androidx.activity.enableEdgeToEdge
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.Scaffold
 import androidx.compose.ui.Modifier
-import com.example.kuitandroidapiexample.navigation.MainNavHost
+import com.example.kuitandroidapiexample.ui.navigation.MainNavHost
 import com.example.kuitandroidapiexample.ui.theme.KuitAndroidApiExampleTheme
 
 class MainActivity : ComponentActivity() {
