+++ app/src/main/java/com/example/kuitandroidapiexample/ui/navigation/MainNavHost.kt
@@ -1,4 +1,4 @@
-package com.example.kuitandroidapiexample.navigation
+package com.example.kuitandroidapiexample.ui.navigation
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.runtime.Composable
@@ -6,9 +6,9 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.toRoute
-import com.example.kuitandroidapiexample.detail.screen.DetailScreen
-import com.example.kuitandroidapiexample.home.screen.HomeScreen
-import com.example.kuitandroidapiexample.register.screen.RegisterScreen
+import com.example.kuitandroidapiexample.ui.detail.screen.DetailScreen
+import com.example.kuitandroidapiexample.ui.home.screen.HomeScreen
+import com.example.kuitandroidapiexample.ui.register.screen.RegisterScreen
 
 @Composable
 fun MainNavHost(
