+++ app/src/main/java/com/example/kuitandroidapiexample/ui/navigation/Route.kt
@@ -1,4 +1,4 @@
-package com.example.kuitandroidapiexample.navigation
+package com.example.kuitandroidapiexample.ui.navigation
 
 import kotlinx.serialization.Serializable
