+++ app/src/main/java/com/example/kuitandroidapiexample/ui/register/componet/FindUTextField.kt
@@ -1,4 +1,4 @@
-package com.example.kuitandroidapiexample.register.componet
+package com.example.kuitandroidapiexample.ui.register.componet
 
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Column
